register UI, react-router-dom, folder structure
