+++ devstudy/src/pages/Register/index.tsx
@@ -1,0 +1,6 @@
+import Register from "../../component/unit/register/register";
+
+const RegisterPage = () => {
+  return <Register />;
+};
+export default RegisterPage;
