+++ devstudy/src/util/navigation-util.ts
@@ -1,0 +1,3 @@
+export const NavigationUtil = {
+  register: "/register",
+};
